print.summary.voronoi {interp}	R Documentation

Print a summary of a voronoi object

Description

Prints some information about object x

Usage

## S3 method for class 'summary.voronoi'
print(x, ...)

Arguments

x
object of class "summary.voronoi", generated by summary.voronoi.

...
additional paramters for print

Value

None

Note

This function is meant as replacement for the function of same name in package tripack and should be fully backward compatible.

Author(s)

Albrecht Gebhardt <albrecht.gebhardt@aau.at>, Roger Bivand <roger.bivand@nhh.no>

See Also

voronoi,voronoi.mosaic, print.voronoi, plot.voronoi, summary.voronoi.

[Package interp version 1.1-6 Index]