summary.triSht {interp}	R Documentation

Return a summary of a triangulation object

Description

Returns some information (number of nodes, triangles, arcs) about object.

Usage

## S3 method for class 'triSht'
summary(object,...)

Arguments

object
object of class "triSht"

...
additional paramters for summary

Value

An object of class "summary.triSht", to be printed by print.summary.triSht.

It contains the number of nodes (n), of arcs (na), of boundary nodes (nb) and triangles (nt).

Note

This function is meant as replacement for the function of same name in package tripack.

The only difference is that no constraints are possible with triSht objects of package interp.

Author(s)

Albrecht Gebhardt <albrecht.gebhardt@aau.at>, Roger Bivand <roger.bivand@nhh.no>

See Also

triSht, print.triSht, plot.triSht, print.summary.triSht.

[Package interp version 1.1-6 Index]