triSht {interp}	R Documentation

A triangulation object

Description

R object that represents the triangulation of a set of 2D points, generated by tri.mesh.

Arguments

n
Number of nodes

x
x coordinates of the triangulation nodes

y
y coordinates of the triangulation nodes

nt
number of triangles

trlist
Matrix of indices which defines the triangulation, each row corresponds to a triangle.

Columns i1, i2, i3 of the row i contain the node indices defining the ith triangle.

Columns j1, j2, j3 of the row i contain the indices of neighbour triangles (or 0 if no neighbour available along the convex hull).

Columns k1, k2, k3 of the row i contain the indices of the arcs of the ith triangle as returned by the arcs function.

cclist
Matrix describing the circumcircles and triangles.

Columns x and y contain coordinates of the circumcircle centers, r is the circumcircle radius.

area is the triangle area and ratio is the ratio of the radius of the inscribed circle to the circumcircle radius. It takes it maximum value 0.5 for an equilateral triangle.

The radius of the inscribed circle can be get via r_i=\frac{r}{ratio}.

nchull
number of points on the convex hull

chull
A vector containing the indices of nodes forming the convec hull (in counterclockwise ordering).

narcs
number of arcs forming the triangulation

arcs
A matrix with node indices describing the arcs, contains two columns from and to.

call
call, which generated this object

Note

This object is not backward compatible with tri objects generated from package tripack but the functions and methods are! So you have to regenerate these objects and then you can continue to use the same calls as before.

The only difference is that no constraints to the triangulation are possible in package interp.

Function triSht2tri provides an option to convert this object into the older form from package tripack, but it will not generate exact copies as if the object would have been created with tripack::tri.mesh! The old data structure consists of three lists describing adjacency lists of triangulation nodes in counterclockwise order, the translation function only genrates such a valid (but not unique) description.

Author(s)

Albrecht Gebhardt <albrecht.gebhardt@aau.at>, Roger Bivand <roger.bivand@nhh.no>

See Also

tri.mesh, print.triSht,triSht2tri, plot.triSht, summary.triSht

[Package interp version 1.1-6 Index]