voronoi {interp}	R Documentation

Voronoi object

Description

A voronoi object is created with voronoi.mosaic

Arguments

x,y
x and y coordinates of nodes of the voronoi mosaic. Each node is a circumcircle center of some triangle from the Delaunay triangulation.

node
logical vector, indicating real nodes of the voronoi mosaic. These nodes are the centers of circumcircles of triangles with positive area of the delaunay triangulation.

If node[i]=FALSE, (c[i],x[i]) belongs to a triangle with area 0.

n1,n2,n3
indices of neighbour nodes. Negative indices indicate dummy points as neighbours.

tri
triangulation object, see triSht.

area
area of triangle i.

ratio
aspect ratio (inscribed radius/circumradius) of triangle i.

radius
circumradius of triangle i.

dummy.x,dummy.y
x and y coordinates of dummy points. They are used for plotting of unbounded tiles.

Note

This version of voronoi object is generated from the tri.mesh function from package interp. That's the only difference to voronoi objects generated with package tripack.

Author(s)

Albrecht Gebhardt <albrecht.gebhardt@aau.at>, Roger Bivand <roger.bivand@nhh.no>

See Also

voronoi.mosaic,plot.voronoi

[Package interp version 1.1-6 Index]