voronoi.findrejectsites {interp}	R Documentation

Find the Voronoi sites at the border of the region (to be rejected).

Description

Find the sites in the Voronoi tesselation that lie at the edge of the region. A site is at the edge if any of the vertices of its Voronoi polygon lie outside the rectangle with corners (xmin,ymin) and (xmax,ymax).

Usage

voronoi.findrejectsites(voronoi.obj, xmin, xmax, ymin, ymax)

Arguments

voronoi.obj
object of class "voronoi"

xmin
minimum x-coordinate of sites in the region

xmax
maximum x-coordinate of sites in the region

ymin
minimum y-coordinate of sites in the region

ymax
maximum y-coordinate of sites in the region

Value

A logical vector of the same length as the number of sites. If the site is a reject, the corresponding element of the vector is set to TRUE.

Author(s)

S. J. Eglen

See Also

voronoi.polygons

[Package interp version 1.1-6 Index]